refactor: Extract duplicated test case
