+++ src/test/java/gift/option/OptionControllerTest.java
@@ -18,16 +18,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import gift.option.OptionTestCase.OptionNameAllowedCharacterError;
+import gift.option.OptionTestCase.OptionNameLengthError;
+import gift.option.OptionTestCase.OptionQuantitySizeError;
 import gift.token.JwtProvider;
 import java.util.List;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -48,43 +47,6 @@ public class OptionControllerTest {
 
     private static final String URL = "/api/products/%d/option";
     private final ObjectMapper objectMapper = new ObjectMapper();
-
-    static class OptionQuantitySizeError implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of(0),
-                Arguments.of(100000000),
-                Arguments.of(Integer.MAX_VALUE),
-                Arguments.of(Integer.MIN_VALUE)
-            );
-        }
-    }
-
-    static class OptionNameLengthErrorMethodSource implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of("thisSentenceIsTooLongUseForOptionNameDoesNotItHelloWorldHelloWorld"),
-                Arguments.of("이문장은옵션의이름으로쓰기에는너무길어요안그런가요50자나되는문장을뭐라고써야할지도모르겠네요이제는딱맞는거같아요"),
-                Arguments.of(
-                    "공백 포함 공백 포함 공백 포함 공백 포함 [] 공백 포함 [ ] ( ) + - & / _ 공백 포함 공백 포함 공백 포함 ")
-            );
-        }
-    }
-
-    static class OptionNameAllowedCharacterMethodSource implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of("한글과영어 그리고 특수문자 ()[]+-&/_ 😀"),
-                Arguments.of("~!@#$%^&*()_+{}|\"'")
-            );
-        }
-    }
 
     @Nested
     @DisplayName("[Unit] get option test")
@@ -195,7 +157,7 @@ public class OptionControllerTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameAllowedCharacterMethodSource.class)
+        @ArgumentsSource(OptionNameAllowedCharacterError.class)
         @DisplayName("option name allowed character error")
         void optionNameAllowedCharacterError(String optionName) throws Exception {
             //given
@@ -216,7 +178,7 @@ public class OptionControllerTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameLengthErrorMethodSource.class)
+        @ArgumentsSource(OptionNameLengthError.class)
         @DisplayName("option name length error")
         void optionNameLengthError(String optionName) throws Exception {
             //given
@@ -325,7 +287,7 @@ public class OptionControllerTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameAllowedCharacterMethodSource.class)
+        @ArgumentsSource(OptionNameAllowedCharacterError.class)
         @DisplayName("option name allowed character error")
         void optionNameAllowedCharacterError(String optionName) throws Exception {
             //given
@@ -346,7 +308,7 @@ public class OptionControllerTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameLengthErrorMethodSource.class)
+        @ArgumentsSource(OptionNameLengthError.class)
         @DisplayName("option name length error")
         void optionNameLengthError(String optionName) throws Exception {
             //given
