+++ src/test/java/gift/option/OptionServiceTest.java
@@ -16,19 +16,18 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import gift.category.Category;
+import gift.option.OptionTestCase.OptionNameAllowedCharacterError;
+import gift.option.OptionTestCase.OptionNameLengthError;
+import gift.option.OptionTestCase.OptionQuantitySizeError;
 import gift.product.Product;
 import gift.product.ProductRepository;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -51,43 +50,6 @@ public class OptionServiceTest {
     private Product product;
     private OptionDTO optionDTO;
     private Option option;
-
-    static class OptionQuantitySizeError implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of(0),
-                Arguments.of(100000000),
-                Arguments.of(Integer.MAX_VALUE),
-                Arguments.of(Integer.MIN_VALUE)
-            );
-        }
-    }
-
-    static class OptionNameLengthErrorMethodSource implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of("thisSentenceIsTooLongUseForOptionNameDoesNotItHelloWorldHelloWorld"),
-                Arguments.of("이문장은옵션의이름으로쓰기에는너무길어요안그런가요50자나되는문장을뭐라고써야할지도모르겠네요이제는딱맞는거같아요"),
-                Arguments.of(
-                    "공백 포함 공백 포함 공백 포함 공백 포함 [] 공백 포함 [ ] ( ) + - & / _ 공백 포함 공백 포함 공백 포함 ")
-            );
-        }
-    }
-
-    static class OptionNameAllowedCharacterMethodSource implements ArgumentsProvider {
-
-        @Override
-        public Stream<Arguments> provideArguments(ExtensionContext context) {
-            return Stream.of(
-                Arguments.of("한글과영어 그리고 특수문자 ()[]+-&/_ 😀"),
-                Arguments.of("~!@#$%^&*()_+{}|\"'")
-            );
-        }
-    }
 
     @BeforeEach
     void setUp() {
@@ -223,7 +185,7 @@ public class OptionServiceTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameAllowedCharacterMethodSource.class)
+        @ArgumentsSource(OptionNameAllowedCharacterError.class)
         @DisplayName("option name allowed character error")
         void optionNameAllowedCharacterError(String optionName) {
             //given
@@ -240,7 +202,7 @@ public class OptionServiceTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameLengthErrorMethodSource.class)
+        @ArgumentsSource(OptionNameLengthError.class)
         @DisplayName("option name length error")
         void optionNameLengthError(String optionName) {
             //given
@@ -355,7 +317,7 @@ public class OptionServiceTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameAllowedCharacterMethodSource.class)
+        @ArgumentsSource(OptionNameAllowedCharacterError.class)
         @DisplayName("option name allowed character error")
         void optionNameAllowedCharacterError(String optionName) {
             //given
@@ -372,7 +334,7 @@ public class OptionServiceTest {
         }
 
         @ParameterizedTest
-        @ArgumentsSource(OptionNameLengthErrorMethodSource.class)
+        @ArgumentsSource(OptionNameLengthError.class)
         @DisplayName("option name length error")
         void optionNameLengthError(String optionName) {
             //given
